requirements.txt
# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# WebSocket support
python-socketio>=5.8.0
eventlet>=0.33.0

# Database
sqlalchemy>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
coverage>=7.3.0

# Code quality
flake8>=6.1.0
black>=23.9.0
